pyproject.toml
[project]
name = "spotify-money-calculator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiocache[redis]>=0.12.3",
    "dotenv>=0.9.9",
    "fastapi>=0.123.5",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pusher>=3.3.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python3-discogs-client>=2.8",
    "ruff>=0.14.7",
    "uvicorn>=0.38.0",
]
